耕地可选择水源限制，添加悦灵喂养，雕刻南瓜，瓶子草和火把花催熟收割

+++ src/main/java/com/flier268/autoharvest/TaskManager.java
@@ -29,7 +29,7 @@ public class TaskManager {
         taskList.add(new Line(Commands.MOVEITEM, slotNumber, currentHotbarSlot));
     }
 
-    public void Add_TickSkip(float skipTick) {
+    public void Add_TickSkip(int skipTick) {
         for (int i = 0; i < skipTick; i++)
             taskList.add(new Line(Commands.SKIPTICK));
     }
